backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httpx==0.25.2
websockets==12.0
python-socketio==5.10.0
bcrypt==4.1.2
email-validator==2.1.0
pillow==10.4.0
dnspython==2.4.2
python-dateutil==2.8.2
google-generativeai>=0.8.0
